+++ MathTest/Matrix.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "pch.h"
 
-Matrix CreateMatrix2x2(double, double, double, double);
-Matrix CreateMatrix2x3(double, double, double, double, double, double);
-Matrix CreateMatrix3x2(double, double, double, double, double, double);
-Matrix CreateMatrix3x3(double, double, double, double, double, double, double, double, double);
+Matrix CreateMatrix2x2(float, float, float, float);
+Matrix CreateMatrix2x3(float, float, float, float, float, float);
+Matrix CreateMatrix3x2(float, float, float, float, float, float);
+Matrix CreateMatrix3x3(float, float, float, float, float, float, float, float, float);
